backend/requirements.txt
Flask-SocketIO
google-cloud-speech~=2.15
